fix merge: adjust set_tokenizer that was moved
