+++ dataquality/core/log.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Tuple, Union
-from warnings import warn
 
 import numpy as np
 import pandas as pd
@@ -605,51 +604,6 @@ def set_epoch_and_split(
     """
     set_epoch(epoch)
     set_split(split, inference_name)
-
-
-@check_noop
-def set_tokenizer(
-    tokenizer: PreTrainedTokenizerFast,
-    max_input_tokens: Optional[int] = None,
-    max_target_tokens: Optional[int] = None,
-) -> None:
-    """Seq2seq only. Set the tokenizer for your run
-
-    Must be a fast tokenizer, and must support `decode`, `encode`, `encode_plus`
-    """
-    task_type = get_task_type()
-    assert task_type == TaskType.seq2seq, "This method is only supported for seq2seq"
-    assert isinstance(
-        tokenizer, PreTrainedTokenizerFast
-    ), "Tokenizer must be an instance of PreTrainedTokenizerFast"
-    assert getattr(tokenizer, "is_fast", False), "Tokenizer must be a fast tokenizer"
-    for attr in ["encode", "decode", "encode_plus", "padding_side"]:
-        assert hasattr(tokenizer, attr), f"Tokenizer must support `{attr}`"
-    seq2seq_logger_config.tokenizer = tokenizer
-
-    seq2seq_logger_config.max_input_tokens = max_input_tokens
-    if seq2seq_logger_config.max_input_tokens is None:
-        seq2seq_logger_config.max_input_tokens = tokenizer.model_max_length
-        warn(
-            (
-                "The argument max_input_tokens is not set, we will use the value "
-                "found in tokenizer.model_max_length. If you tokenized the input with "
-                "another value, this can lead to confusing insights about this "
-                "training run."
-            )
-        )
-
-    seq2seq_logger_config.max_target_tokens = max_target_tokens
-    if seq2seq_logger_config.max_target_tokens is None:
-        warn(
-            (
-                "The argument max_target_tokens is not set. Passing it as an argument "
-                "will speed up logging of the dataset."
-            )
-        )
-
-    # Seq2Seq doesn't have labels but we need to set this to avoid validation errors
-    seq2seq_logger_config.labels = []
 
 
 @check_noop
